+++ Sources/SuperwallKit/Dependencies/DependencyContainer.swift
@@ -69,7 +69,8 @@ final class DependencyContainer {
     receiptManager = ReceiptManager(
       storeKitVersion: options.storeKitVersion,
       productsManager: productsManager,
-      receiptDelegate: purchaseController as? ReceiptDelegate
+      receiptDelegate: purchaseController as? ReceiptDelegate,
+      factory: self
     )
 
     webEntitlementRedeemer = WebEntitlementRedeemer(
@@ -530,8 +531,8 @@ extension DependencyContainer: UserAttributesPlacementFactory {
 
 // MARK: - Receipt Factory
 extension DependencyContainer: ReceiptFactory {
-  func loadPurchasedProducts() async {
-    await receiptManager.loadPurchasedProducts()
+  func loadPurchasedProducts(config: Config? = nil) async {
+    await receiptManager.loadPurchasedProducts(config: config)
   }
 
   func refreshSK1Receipt() async {